frontend-desktop/requirements.txt
# ParkAI Desktop - Python Dependencies
# Compatible with Raspberry Pi OS

# ⚠️ IMPORTANT for Raspberry Pi users:
# PyQt6 MUST be installed via apt (system package), NOT pip!
# Run: sudo apt install python3-pyqt6 python3-pyqt6.qtwebengine
# Then create venv with: python3 -m venv --system-site-packages venv

# Or simply run: make setup (handles everything automatically)

# ===== Core Dependencies (via pip) =====

# HTTP client
requests>=2.31.0

# WebSocket client
websocket-client>=1.7.0

# Environment variables
python-dotenv>=1.0.0

# WebRTC support
aiortc>=1.6.0
av>=11.0.0

# Image processing & video optimization
opencv-python>=4.8.0
numpy>=1.24.0,<1.27.0  # Compatible with scipy 1.10.1 (required by aiortc/av)

# ===== Optional Dependencies =====

# Pillow>=10.0.0

# Charts/graphs (nếu muốn thêm analytics)
# pyqtgraph>=0.13.0
# matplotlib>=3.7.0

# ===== System Packages Required (install via apt) =====
# These are NOT in pip, install via apt-get:
# - python3-pyqt6              (PyQt6 GUI framework)
# - python3-pyqt6.qtwebengine  (WebEngine component)
# - python3-dev                (Python development headers)
# - libxcb-xinerama0           (X11 library)
# - libxcb-cursor0             (X11 cursor library)
# - libegl1                    (OpenGL library)
# - libgl1-mesa-glx            (Mesa OpenGL library)

# Run this command on Raspberry Pi to install all system deps:
# sudo apt install python3-pyqt6 python3-pyqt6.qtwebengine python3-dev \
#   libxcb-xinerama0 libxcb-cursor0 libegl1 libgl1-mesa-glx
